requirements.txt
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
h11==0.14.0
idna==3.4
numpy==1.24.2
outcome==1.2.0
pandas==1.5.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
python-dotenv~=1.0.0